Extracted Product Price Service from Product Service
