+++ src/main/java/com/myretail/api/service/ProductService.java
@@ -16,12 +16,4 @@ public interface ProductService {
      */
     ProductDTO findById(int id) throws ProductNotFoundException;
 
-    /**
-     * Updates a product price in the database.
-     * This method is not responsible for creation of new productPrices.
-     * @param productPriceDTO pass the entire object to update a matching productPrice
-     * @throws ProductNotFoundException if the product id doesn't match any productPrice id in the database.
-     */
-    void updateById(ProductPriceDTO productPriceDTO) throws ProductNotFoundException;
-
 }
